init feature example, not tests yet
