+++ bases/bot_detector/api_public/src/api/v2/__init__.py
@@ -1,8 +1,9 @@
+from bot_detector.features.public_api.feedback.routes import router as feedback_router
 from fastapi import APIRouter
 
-from . import feedback, player, report
+from . import player, report
 
 router = APIRouter()
 router.include_router(player.router)
 router.include_router(report.router)
-router.include_router(feedback.router)
+router.include_router(feedback_router)
