+++ bases/bot_detector/api_public/src/api/v2/report.py
@@ -2,9 +2,9 @@ import logging
 
 from bot_detector.api_public.src.app.repositories.player import Player
 from bot_detector.api_public.src.app.repositories.report import CustomError, Report
-from bot_detector.api_public.src.app.views.response.ok import Ok
 from bot_detector.api_public.src.core._cache import SimpleALRUCache
 from bot_detector.api_public.src.core.fastapi.dependencies.session import get_session
+from bot_detector.features.public_api.response.schemas import Ok
 from bot_detector.structs import Detection, ParsedDetection
 from fastapi import APIRouter, Depends, status
 from fastapi.exceptions import HTTPException
