+++ test/__init__
@@ -1,0 +1,3 @@
+from dotenv import find_dotenv, load_dotenv
+
+load_dotenv(find_dotenv())
