+++ test/bases/bot_detector/api_public/test_feedback.py
@@ -1,0 +1,7 @@
+# test_feedback.py
+
+import pytest
+from bot_detector.features.public_api.feedback.routes import Feedback
+
+# @pytest.mark.asyncio
+# async def test_post_feedback_returns_201_when_valid():
